Move helper functions to submarinerbrokerinfo
These are only used in this module and don't need to be lumped into a
general helpers module:

- GetIPSecPSK
- GetBrokerAPIServer
- GetBrokerTokenAndCA

This also eliminates duplication of unit tests.
