+++ pkg/helpers/helpers_test.go
@@ -20,13 +20,11 @@ import (
 
 	"github.com/openshift/library-go/pkg/operator/events/eventstesting"
 
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/diff"
-	dynamicfake "k8s.io/client-go/dynamic/fake"
 	kubefake "k8s.io/client-go/kubernetes/fake"
 	clienttesting "k8s.io/client-go/testing"
 )
@@ -212,304 +210,6 @@ func TestUpdateManagedClusterAddOnStatus(t *testing.T) {
 				if !equality.Semantic.DeepEqual(expected, actual) {
 					t.Errorf(diff.ObjectDiff(expected, actual))
 				}
-			}
-		})
-	}
-}
-
-func TestGenerateAndGetIPSecPSKSecret(t *testing.T) {
-	cases := []struct {
-		name            string
-		brokerNamespace string
-		existings       []runtime.Object
-		expectErr       bool
-	}{
-		{
-			name:            "exist secret",
-			brokerNamespace: "cluster1-broker",
-			existings: []runtime.Object{
-				&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      IPSecPSKSecretName,
-						Namespace: "cluster1-broker",
-					},
-					Data: map[string][]byte{
-						"psk": []byte("abcd1234"),
-					},
-				},
-			},
-			expectErr: false,
-		},
-		{
-			name:            "no secret",
-			brokerNamespace: "cluster2-broker",
-			existings:       []runtime.Object{},
-			expectErr:       false,
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			fakeClient := kubefake.NewSimpleClientset(c.existings...)
-			err := GenerateIPSecPSKSecret(fakeClient, c.brokerNamespace)
-			if err != nil && !c.expectErr {
-				t.Errorf("expect no err: %+v", err)
-			}
-			if err == nil && c.expectErr {
-				t.Errorf("expect err")
-			}
-
-			ipSecPSK, err := GetIPSecPSK(fakeClient, c.brokerNamespace)
-			if err != nil && !c.expectErr {
-				t.Errorf("expect no err: %+v", err)
-			}
-			if err == nil && c.expectErr {
-				t.Errorf("expect err")
-			}
-			if ipSecPSK == "" && !c.expectErr {
-				t.Errorf("expect an valid ipSecPSK")
-			}
-		})
-	}
-}
-
-func TestGetBrokerAPIServer(t *testing.T) {
-	cases := []struct {
-		name      string
-		existings []runtime.Object
-		expectErr bool
-	}{
-		{
-			name:      "no infrastructure",
-			existings: []runtime.Object{},
-			expectErr: true,
-		},
-		{
-			name: "exist infrastructure",
-			existings: []runtime.Object{
-				&unstructured.Unstructured{
-					Object: map[string]interface{}{
-						"apiVersion": "config.openshift.io/v1",
-						"kind":       "Infrastructure",
-						"metadata": map[string]interface{}{
-							"name": "cluster",
-						},
-						"status": map[string]interface{}{
-							"apiServerURL": "https://api.test.dev04.red-chesterfield.com:6443",
-						},
-					},
-				},
-			},
-			expectErr: false,
-		},
-		{
-			name: "no apiServerURL in infrastructure status",
-			existings: []runtime.Object{
-				&unstructured.Unstructured{
-					Object: map[string]interface{}{
-						"apiVersion": "config.openshift.io/v1",
-						"kind":       "Infrastructure",
-						"metadata": map[string]interface{}{
-							"name": "cluster",
-						},
-					},
-				},
-			},
-			expectErr: true,
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			fakeClient := dynamicfake.NewSimpleDynamicClient(runtime.NewScheme(), c.existings...)
-			apiServer, err := GetBrokerAPIServer(fakeClient)
-			if c.expectErr && err == nil {
-				t.Errorf("expect err")
-			}
-			if !c.expectErr && err != nil {
-				t.Errorf("expect no err: %v", err)
-			}
-			if !c.expectErr && err == nil && len(apiServer) == 0 {
-				t.Errorf("expect valid apiServer")
-			}
-		})
-	}
-}
-
-func TestGetBrokerTokenAndCA(t *testing.T) {
-	os.Setenv("BROKER_API_SERVER", "127.0.0.1:6443")
-	defer os.Unsetenv("BROKER_API_SERVER")
-
-	cases := []struct {
-		name            string
-		brokerNamespace string
-		clusterName     string
-		kubeObjs        []runtime.Object
-		ocpObjs         []runtime.Object
-		expectErr       bool
-	}{
-		{
-			name:            "no sa no secret",
-			brokerNamespace: "cluster1-broker",
-			clusterName:     "cluster1",
-			kubeObjs:        []runtime.Object{},
-			ocpObjs:         []runtime.Object{},
-			expectErr:       true,
-		},
-		{
-			name:            "exist sa no secret",
-			brokerNamespace: "cluster1-broker",
-			clusterName:     "cluster1",
-			kubeObjs: []runtime.Object{
-				&corev1.ServiceAccount{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1",
-						Namespace: "cluster1-broker",
-					},
-				},
-			},
-			ocpObjs:   []runtime.Object{},
-			expectErr: true,
-		},
-		{
-			name:            "exist sa and secret, but the secret cannot be found",
-			brokerNamespace: "cluster1-broker",
-			clusterName:     "cluster1",
-			kubeObjs: []runtime.Object{
-				&corev1.ServiceAccount{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1",
-						Namespace: "cluster1-broker",
-					},
-					Secrets: []corev1.ObjectReference{{Name: "cluster-cluster1-token-5pw5c"}},
-				},
-				&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster-cluster1-token-5pw5c",
-						Namespace: "cluster1-broker",
-					},
-					Data: map[string][]byte{
-						"ca.crt": []byte("ca"),
-						"token":  []byte("token"),
-					},
-					Type: corev1.SecretTypeServiceAccountToken,
-				},
-			},
-			ocpObjs:   []runtime.Object{},
-			expectErr: true,
-		},
-		{
-			name:            "exist sa and secret",
-			brokerNamespace: "cluster1-broker",
-			clusterName:     "cluster1",
-			kubeObjs: []runtime.Object{
-				&corev1.ServiceAccount{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1",
-						Namespace: "cluster1-broker",
-					},
-					Secrets: []corev1.ObjectReference{{Name: "cluster1-token-5pw5c"}},
-				},
-				&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1-token-5pw5c",
-						Namespace: "cluster1-broker",
-					},
-					Data: map[string][]byte{
-						"ca.crt": []byte("ca"),
-						"token":  []byte("token"),
-					},
-					Type: corev1.SecretTypeServiceAccountToken,
-				},
-			},
-			ocpObjs:   []runtime.Object{},
-			expectErr: false,
-		},
-		{
-			name:            "get sa from kubeapi server secret",
-			brokerNamespace: "cluster1-broker",
-			clusterName:     "cluster1",
-			kubeObjs: []runtime.Object{
-				&corev1.ServiceAccount{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1",
-						Namespace: "cluster1-broker",
-					},
-					Secrets: []corev1.ObjectReference{{Name: "cluster1-token-5pw5c"}},
-				},
-				&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "cluster1-token-5pw5c",
-						Namespace: "cluster1-broker",
-					},
-					Data: map[string][]byte{
-						"ca.crt": []byte("ca"),
-						"token":  []byte("token"),
-					},
-					Type: corev1.SecretTypeServiceAccountToken,
-				},
-				&corev1.Secret{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "test",
-						Namespace: "openshift-config",
-					},
-					Data: map[string][]byte{
-						"tls.crt": []byte("tls"),
-					},
-					Type: corev1.SecretTypeTLS,
-				},
-			},
-			ocpObjs: []runtime.Object{
-				&unstructured.Unstructured{
-					Object: map[string]interface{}{
-						"apiVersion": "config.openshift.io/v1",
-						"kind":       "Infrastructure",
-						"metadata": map[string]interface{}{
-							"name": "cluster",
-						},
-						"status": map[string]interface{}{
-							"apiServerURL": "https://api.test.dev04.red-chesterfield.com:6443",
-						},
-					},
-				},
-				&unstructured.Unstructured{
-					Object: map[string]interface{}{
-						"apiVersion": "config.openshift.io/v1",
-						"kind":       "APIServer",
-						"metadata": map[string]interface{}{
-							"name": "cluster",
-						},
-						"spec": map[string]interface{}{
-							"servingCerts": map[string]interface{}{
-								"namedCertificates": []interface{}{
-									map[string]interface{}{
-										"names": []interface{}{"api.test.dev04.red-chesterfield.com"},
-										"servingCertificate": map[string]interface{}{
-											"name": "test",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectErr: false,
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			fakeKubeClient := kubefake.NewSimpleClientset(c.kubeObjs...)
-			fakeDynamicClient := dynamicfake.NewSimpleDynamicClient(runtime.NewScheme(), c.ocpObjs...)
-			token, ca, err := GetBrokerTokenAndCA(fakeKubeClient, fakeDynamicClient, c.brokerNamespace, c.clusterName)
-			if err != nil && !c.expectErr {
-				t.Errorf("expect no err: %+v", err)
-			}
-			if err == nil && c.expectErr {
-				t.Errorf("expect err")
-			}
-			if (token == "" || ca == "") && !c.expectErr {
-				t.Errorf("expect valid token and ca")
 			}
 		})
 	}
